+++ ddd/editing.C
@@ -229,7 +229,7 @@ static void isearch_again(ISearchState new_isearch_state, XEvent *event)
 	if (history < 0)
 	    XtCallActionProc(gdb_w, "beep", event, 0, 0);
 	else
-	    isearch_done(XtPointer(history), 0);
+	    isearch_done(XtPointer(intptr_t(history)), 0);
     }
     else
     {
@@ -352,7 +352,7 @@ static bool do_isearch(Widget, XmTextVerifyCallbackStruct *change)
 
 	// Redraw current line with appropriate prompt.
 	XtAppAddTimeOut(XtWidgetToApplicationContext(gdb_w), 0, 
-			isearch_done, XtPointer(history));
+			isearch_done, XtPointer(intptr_t(history)));
 
 	// Upon the next call to gdbMotionCB(), clear ISearch mode,
 	// unless it immediately follows this one.
